+++ tests/HelperFileTest.php
@@ -11,7 +11,7 @@
 
 namespace Tests;
 
-use Gyselroth\Helper\HelperConstantsFile;
+use Gyselroth\Helper\HelperFile;
 use PHPUnit\Framework\Constraint\IsType;
 
 class HelperFileTest extends HelperTestCase
@@ -21,7 +21,7 @@ class HelperFileTest extends HelperTestCase
 
     protected function setUp()
     {
-        $this->uploadedFilePath = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . '01.pdf';
+        $this->uploadedFilePath = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . '01.pdf';
         $pathFileTemplate  = __DIR__ . '/Fixtures/data/files/zip/01.pdf';
         copy($pathFileTemplate, $this->uploadedFilePath);
         $this->uploadedFileInfo = [
@@ -43,7 +43,7 @@ class HelperFileTest extends HelperTestCase
     public function testGetMimes(): void
     {
         $this->assertThat(
-            HelperConstantsFile::MIMES,
+            HelperFile::MIMES,
             new IsType('array')
         );
     }
@@ -53,7 +53,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testGetUniqueFilename(): void
     {
-        $this->assertRegExp('/\d{8}-\d{6}[a-z0-9]{13}/', HelperConstantsFile::getUniqueFilename());
+        $this->assertRegExp('/\d{8}-\d{6}[a-z0-9]{13}/', HelperFile::getUniqueFilename());
     }
 
     /**
@@ -61,7 +61,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testGetUniqueFilenameNoPrefix(): void
     {
-        $this->assertRegExp('/\d{8}-\d{6}/', HelperConstantsFile::getUniqueFilename('', true, false));
+        $this->assertRegExp('/\d{8}-\d{6}/', HelperFile::getUniqueFilename('', true, false));
     }
 
     /**
@@ -69,7 +69,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testGetUniqueFilenameLeadString(): void
     {
-        $this->assertRegExp('/test\d{8}-\d{6}[a-z0-9]{13}/', HelperConstantsFile::getUniqueFilename('test'));
+        $this->assertRegExp('/test\d{8}-\d{6}[a-z0-9]{13}/', HelperFile::getUniqueFilename('test'));
     }
 
     /**
@@ -77,7 +77,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testGetUniqueFilenameWithFileEndingNoDatePrefix(): void
     {
-        $this->assertRegExp('/[a-z0-9]{13}\.zip/', HelperConstantsFile::getUniqueFilename('', false, true, 'zip'));
+        $this->assertRegExp('/[a-z0-9]{13}\.zip/', HelperFile::getUniqueFilename('', false, true, 'zip'));
     }
 
     public function testEnsureFilenamesStartWithPath(): void
@@ -85,7 +85,7 @@ class HelperFileTest extends HelperTestCase
         $files =            ['/test/path/to/file.zip','file','file.png'];
         $filesExpected =    ['/test/path/to/file.zip','/test/path/to/file','/test/path/to/file.png'];
         $path =             '/test/path/to';
-        $this->assertEquals($filesExpected, HelperConstantsFile::ensureFilenamesStartWithPath($files, $path));
+        $this->assertEquals($filesExpected, HelperFile::ensureFilenamesStartWithPath($files, $path));
     }
 
     /**
@@ -93,7 +93,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testValidateFilenameWithPath(): void
     {
-        $this->assertSame('testfile.png', HelperConstantsFile::validateFilename('test/file.png'));
+        $this->assertSame('testfile.png', HelperFile::validateFilename('test/file.png'));
     }
 
     /**
@@ -102,7 +102,7 @@ class HelperFileTest extends HelperTestCase
     public function testValidateFilenameForbiddenCharEqualReplacements(): void
     {
         $this->assertSame('long filename.png',
-            HelperConstantsFile::validateFilename('long2fil3n *ame.png', ['2', '3', ' *'], [' ', 'e', '']));
+            HelperFile::validateFilename('long2fil3n *ame.png', ['2', '3', ' *'], [' ', 'e', '']));
     }
 
     /**
@@ -111,7 +111,7 @@ class HelperFileTest extends HelperTestCase
     public function testValidateFilenameForbiddenCharUnequalReplacements(): void
     {
         $this->assertSame('long filename.png',
-            HelperConstantsFile::validateFilename('long filename.png', [' ', 'e'], ['']));
+            HelperFile::validateFilename('long filename.png', [' ', 'e'], ['']));
     }
 
     /**
@@ -120,7 +120,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testValidateFilenameEmpty(): void
     {
-        HelperConstantsFile::validateFilename('');
+        HelperFile::validateFilename('');
     }
 
     /**
@@ -129,9 +129,9 @@ class HelperFileTest extends HelperTestCase
      */
     public function testWrite(): void
     {
-        $pathFile = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.txt';
+        $pathFile = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.txt';
         $pathFileExpected  = __DIR__ . '/Fixtures/data/files/test.txt';
-        HelperConstantsFile::write($pathFile, 'test');
+        HelperFile::write($pathFile, 'test');
         $this->assertFileEquals($pathFileExpected, $pathFile);
         if (is_file($pathFile)) {
             unlink($pathFile);
@@ -144,10 +144,10 @@ class HelperFileTest extends HelperTestCase
      */
     public function testWriteAppendMode(): void
     {
-        $pathFile = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.txt';
+        $pathFile = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.txt';
         $pathFileTemplate  = __DIR__ . '/Fixtures/data/files/test.txt';
         copy($pathFileTemplate, $pathFile);
-        HelperConstantsFile::write($pathFile, 'file', 'a');
+        HelperFile::write($pathFile, 'file', 'a');
         $this->assertStringEqualsFile($pathFile, 'testfile');
         if (is_file($pathFile)) {
             unlink($pathFile);
@@ -160,12 +160,12 @@ class HelperFileTest extends HelperTestCase
      */
     public function testWriteJson(): void
     {
-        $pathFile = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.json';
+        $pathFile = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.json';
         $array = [
             'testkey' => 'value',
             'testkey2' => 'value2'
         ];
-        HelperConstantsFile::writeJson($pathFile, $array);
+        HelperFile::writeJson($pathFile, $array);
         $this->assertStringEqualsFile($pathFile, '{"testkey":"value","testkey2":"value2"}');
         if (is_file($pathFile)) {
             unlink($pathFile);
@@ -174,14 +174,14 @@ class HelperFileTest extends HelperTestCase
 
     public function testWriteCsv(): void
     {
-        $pathFile = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.csv';
+        $pathFile = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'test.csv';
         $header =
             ['Name',  'Beschäftigung', 'Firma'];
         $data = [
             ['Kay',   'Fulltime',      'Gyselroth'],
             ['Ewald', 'Parttime',      'Gyselroth']
         ];
-        HelperConstantsFile::writeCsv($data, $header, $pathFile);
+        HelperFile::writeCsv($data, $header, $pathFile);
         $this->assertStringEqualsFile($pathFile, "Name,Beschäftigung,Firma\nKay,Fulltime,Gyselroth\nEwald,Parttime,Gyselroth\n", 'Writing a csv file appends an empty new line.');
         if (is_file($pathFile)) {
             unlink($pathFile);
@@ -191,7 +191,7 @@ class HelperFileTest extends HelperTestCase
     public function testScanDir(): void
     {
         $path = __DIR__ . '/Fixtures/data/files/unzip';
-        $array = HelperConstantsFile::scanDir($path);
+        $array = HelperFile::scanDir($path);
         $this->assertCount(2, $array);
         $this->assertStringEndsWith('/data/files/unzip/tobeunzipped.zip', $array[0]);
         $this->assertStringEndsWith('/data/files/unzip/unzipped', $array[1]);
@@ -200,7 +200,7 @@ class HelperFileTest extends HelperTestCase
     public function testScanDirFilter(): void
     {
         $path = __DIR__ . '/Fixtures/data/files/unzip';
-        $array = HelperConstantsFile::scanDir($path, 'zip');
+        $array = HelperFile::scanDir($path, 'zip');
         $this->assertCount(1, $array);
         $this->assertStringEndsWith('/data/files/unzip/tobeunzipped.zip', $array[0]);
     }
@@ -208,7 +208,7 @@ class HelperFileTest extends HelperTestCase
     public function testScanDirRecursive(): void
     {
         $path = __DIR__ . '/Fixtures/data/files/unzip';
-        $array = HelperConstantsFile::scanDir($path, '', true);
+        $array = HelperFile::scanDir($path, '', true);
         $this->assertCount(10, $array);
         $this->assertStringEndsWith('/data/files/unzip/unzipped/01.pdf', $array[7]);
         $this->assertStringEndsWith('/data/files/unzip/tobeunzipped.zip', $array[9]);
@@ -222,20 +222,20 @@ class HelperFileTest extends HelperTestCase
     public function testChmodRecursive(): void
     {
         //...
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'folder';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'folder';
         if (!is_dir($path)) {
             mkdir($path);
         }
         $file = $path . DIRECTORY_SEPARATOR . '02.pdf';
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::chmodRecursive($path);
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::chmodRecursive($path);
         $file
         ? $this->assertFileIsWritable($file)
         : $this->assertFileNotIsWritable($file);
-        HelperConstantsFile::chmodRecursive($path, '0600');
+        HelperFile::chmodRecursive($path, '0600');
         if (is_dir($path)) {
-            HelperConstantsFile::rmdirRecursive($path);
+            HelperFile::rmdirRecursive($path);
         }
     }
 
@@ -248,7 +248,7 @@ class HelperFileTest extends HelperTestCase
     {
         $files = ['intranet.test.js', 'intranet.test.sub.js', 'intranet.js', 'intranet.othertest.js'];
         $sortedFiles = ['intranet.js', 'intranet.othertest.js', 'intranet.test.js', 'intranet.test.sub.js'];
-        $this->assertEquals($sortedFiles, HelperConstantsFile::sortByDepth($files));
+        $this->assertEquals($sortedFiles, HelperFile::sortByDepth($files));
     }
 
     public function testGetDirectoryInfo(): void
@@ -258,7 +258,7 @@ class HelperFileTest extends HelperTestCase
             'items' => 9,
             'size' => 36864
         ];
-        $this->assertEquals($expected, HelperConstantsFile::getDirectoryInfo($path));
+        $this->assertEquals($expected, HelperFile::getDirectoryInfo($path));
     }
 
     /**
@@ -267,7 +267,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testGetUploadFileInfo(): void
     {
-        $this->assertSame('PDF document, version 1.4', HelperConstantsFile::getUploadFileInfo($this->uploadedFileInfo));
+        $this->assertSame('PDF document, version 1.4', HelperFile::getUploadFileInfo($this->uploadedFileInfo));
     }
 
     /**
@@ -276,7 +276,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testValidateUploadFileOk(): void
     {
-        $this->assertSame('', HelperConstantsFile::validateUploadFile($this->uploadedFileInfo, ['application/pdf']));
+        $this->assertSame('', HelperFile::validateUploadFile($this->uploadedFileInfo, ['application/pdf']));
     }
 
     /**
@@ -287,7 +287,7 @@ class HelperFileTest extends HelperTestCase
     {
         $uploadedFileInfo = $this->uploadedFileInfo;
         unset($uploadedFileInfo['name']);
-        $this->assertSame('Name der Datei wurde nicht empfangen', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('Name der Datei wurde nicht empfangen', HelperFile::validateUploadFile($uploadedFileInfo, []));
     }
 
     /**
@@ -296,7 +296,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testValidateUploadFileTooLarge(): void
     {
-        $this->assertSame('Die Datei ist zu gross', HelperConstantsFile::validateUploadFile($this->uploadedFileInfo, [], 12091));
+        $this->assertSame('Die Datei ist zu gross', HelperFile::validateUploadFile($this->uploadedFileInfo, [], 12091));
     }
 
     /**
@@ -305,7 +305,7 @@ class HelperFileTest extends HelperTestCase
      */
     public function testValidateUploadFileNotAllowed(): void
     {
-        $this->assertSame('Dateityp ist nicht erlaubt', HelperConstantsFile::validateUploadFile($this->uploadedFileInfo, ['image/png', 'image/jpg']));
+        $this->assertSame('Dateityp ist nicht erlaubt', HelperFile::validateUploadFile($this->uploadedFileInfo, ['image/png', 'image/jpg']));
     }
 
     /**
@@ -316,15 +316,15 @@ class HelperFileTest extends HelperTestCase
     {
         $uploadedFileInfo = $this->uploadedFileInfo;
         $uploadedFileInfo['error'] = UPLOAD_ERR_NO_FILE;
-        $this->assertSame('No file sent.', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('No file sent.', HelperFile::validateUploadFile($uploadedFileInfo, []));
         $uploadedFileInfo['error'] = UPLOAD_ERR_FORM_SIZE;
-        $this->assertSame('Exceeded filesize limit.', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('Exceeded filesize limit.', HelperFile::validateUploadFile($uploadedFileInfo, []));
         $uploadedFileInfo['error'] = UPLOAD_ERR_INI_SIZE;
-        $this->assertSame('Exceeded filesize limit.', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('Exceeded filesize limit.', HelperFile::validateUploadFile($uploadedFileInfo, []));
         $uploadedFileInfo['error'] = 123456789;
-        $this->assertSame('Unknown errors.', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('Unknown errors.', HelperFile::validateUploadFile($uploadedFileInfo, []));
         $uploadedFileInfo['size'] = 0;
-        $this->assertSame('File is empty', HelperConstantsFile::validateUploadFile($uploadedFileInfo, []));
+        $this->assertSame('File is empty', HelperFile::validateUploadFile($uploadedFileInfo, []));
     }
 
     public function testStoreUploadFile(): void
@@ -340,12 +340,12 @@ class HelperFileTest extends HelperTestCase
     public function testUnlinkFiles(): void
     {
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'unlink_folder';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'unlink_folder';
         if (!is_dir($path)) {
             mkdir($path);
         }
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::unlinkFiles($path, ['01.pdf', '03.txt']);
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::unlinkFiles($path, ['01.pdf', '03.txt']);
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '01.pdf');
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '03.txt');
     }
@@ -353,12 +353,12 @@ class HelperFileTest extends HelperTestCase
     public function testDeleteFilesInDirectory(): void
     {
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'delete_folder';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'delete_folder';
         if (!is_dir($path)) {
             mkdir($path);
         }
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::deleteFilesInDirectory($path . DIRECTORY_SEPARATOR . '03.*');
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::deleteFilesInDirectory($path . DIRECTORY_SEPARATOR . '03.*');
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '03.pdf');
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '03.txt');
     }
@@ -369,7 +369,7 @@ class HelperFileTest extends HelperTestCase
     public function testDeleteIfExists()
     {
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'delete_exists';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'delete_exists';
         if (!is_dir($path)) {
             mkdir($path);
         }
@@ -377,15 +377,15 @@ class HelperFileTest extends HelperTestCase
         if (!is_dir($subPath)) {
             mkdir($subPath);
         }
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::copyDirectory($pathCopy, $subPath);
-        $this->assertFalse(HelperConstantsFile::deleteIfExists($path . DIRECTORY_SEPARATOR . 'asdfasdf.pdf'));
-        HelperConstantsFile::deleteIfExists($path . DIRECTORY_SEPARATOR . '01.pdf');
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::copyDirectory($pathCopy, $subPath);
+        $this->assertFalse(HelperFile::deleteIfExists($path . DIRECTORY_SEPARATOR . 'asdfasdf.pdf'));
+        HelperFile::deleteIfExists($path . DIRECTORY_SEPARATOR . '01.pdf');
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '01.pdf');
-        HelperConstantsFile::deleteIfExists(['03.pdf', '02.txt'], $path);
+        HelperFile::deleteIfExists(['03.pdf', '02.txt'], $path);
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '03.pdf');
         $this->assertFileNotExists($path . DIRECTORY_SEPARATOR . '02.txt');
-        HelperConstantsFile::deleteIfExists($path);
+        HelperFile::deleteIfExists($path);
         $this->assertDirectoryNotExists($subPath);
         $this->assertDirectoryNotExists($path);
     }
@@ -393,9 +393,9 @@ class HelperFileTest extends HelperTestCase
     public function testRmdirRecursive(): void
     {
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'remove_recursive';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'remove_recursive';
         $subPath = $path . DIRECTORY_SEPARATOR . 'subfolder';
-        $secondPath = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'remove_recursive2';
+        $secondPath = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'remove_recursive2';
         if (!is_dir($path)) {
             mkdir($path);
         }
@@ -405,9 +405,9 @@ class HelperFileTest extends HelperTestCase
         if (!is_dir($secondPath)) {
             mkdir($secondPath);
         }
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::copyDirectory($pathCopy, $subPath);
-        HelperConstantsFile::rmdirRecursive([$path, $secondPath]);
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::copyDirectory($pathCopy, $subPath);
+        HelperFile::rmdirRecursive([$path, $secondPath]);
         $this->assertDirectoryNotExists($subPath);
         $this->assertDirectoryNotExists($path);
         $this->assertDirectoryNotExists($secondPath);
@@ -431,8 +431,8 @@ class HelperFileTest extends HelperTestCase
      */
     public function testIsDirectoryWritable(): void
     {
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'writable';
-        $this->assertTrue(HelperConstantsFile::isDirectoryWritable($path, false, true));
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'writable';
+        $this->assertTrue(HelperFile::isDirectoryWritable($path, false, true));
         $this->assertDirectoryExists($path);
     }
 
@@ -442,16 +442,16 @@ class HelperFileTest extends HelperTestCase
      */
     public function testIsDirectoryWritableException(): void
     {
-        $this->assertFalse(HelperConstantsFile::isDirectoryWritable('/fasdfas/vggdfgff', true, false));
+        $this->assertFalse(HelperFile::isDirectoryWritable('/fasdfas/vggdfgff', true, false));
     }
 
     public function testSanitizeFilename(): void
     {
         $input  = "thIs__wøn't--BÊ thë_såme \n filename Æfter replaÇing&prÕcessing_thiš...file";
         $output1 = 'this_wont-b-the_same-filename-fter-replaing-and-prcessing_this.file';
-        $this->assertSame($output1, HelperConstantsFile::sanitizeFilename($input));
+        $this->assertSame($output1, HelperFile::sanitizeFilename($input));
         $output2 = 'thIs_wont-BE-the_same-filename-After-replaCing-and-prOcessing_this.file';
-        $this->assertSame($output2, HelperConstantsFile::sanitizeFilename($input, false));
+        $this->assertSame($output2, HelperFile::sanitizeFilename($input, false));
     }
 
     public function testGetBasenames(): void
@@ -476,22 +476,22 @@ class HelperFileTest extends HelperTestCase
             'to_other_file.longfileending',
             'it.pdf'
         ];
-        $this->assertEquals($basenamesUnique, HelperConstantsFile::getBasenames($filePaths));
-        $this->assertEquals($basenames, HelperConstantsFile::getBasenames($filePaths, false));
+        $this->assertEquals($basenamesUnique, HelperFile::getBasenames($filePaths));
+        $this->assertEquals($basenames, HelperFile::getBasenames($filePaths, false));
     }
 
     public function testEnsurePathEndsWithDirectorySeparator(): void
     {
-        $this->assertSame('/this/is/a/path/', HelperConstantsFile::ensurePathEndsWithDirectorySeparator('/this/is/a/path'));
-        $this->assertSame('/this/is/a/path/', HelperConstantsFile::ensurePathEndsWithDirectorySeparator('/this/is/a/path/'));
-        $this->markTestIncomplete('Next assertion skipped: HelperConstantsFile::ensurePathEndsWithDirectorySeparator(\'\') returns \'\' instead of \'/\'');
-        $this->assertSame('/', HelperConstantsFile::ensurePathEndsWithDirectorySeparator(''), 'Should return \'/\'');
+        $this->assertSame('/this/is/a/path/', HelperFile::ensurePathEndsWithDirectorySeparator('/this/is/a/path'));
+        $this->assertSame('/this/is/a/path/', HelperFile::ensurePathEndsWithDirectorySeparator('/this/is/a/path/'));
+        $this->markTestIncomplete('Next assertion skipped: HelperFile::ensurePathEndsWithDirectorySeparator(\'\') returns \'\' instead of \'/\'');
+        $this->assertSame('/', HelperFile::ensurePathEndsWithDirectorySeparator(''), 'Should return \'/\'');
     }
 
     public function testScanFilesystem(): void
     {
         $pathCopy = __DIR__ . '/Fixtures/data/files/unzip/unzipped';
-        $path = HelperConstantsFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'filesystem';
+        $path = HelperFile::getGlobalTmpPath(true) . DIRECTORY_SEPARATOR . 'filesystem';
         if (!is_dir($path)) {
             mkdir($path);
         }
@@ -499,8 +499,8 @@ class HelperFileTest extends HelperTestCase
         if (!is_dir($subPath)) {
             mkdir($subPath);
         }
-        HelperConstantsFile::copyDirectory($pathCopy, $path);
-        HelperConstantsFile::copyDirectory($pathCopy, $subPath);
+        HelperFile::copyDirectory($pathCopy, $path);
+        HelperFile::copyDirectory($pathCopy, $subPath);
         $filesystem = [
             $path . '/01.pdf',
             $path . '/01.txt',
@@ -521,7 +521,7 @@ class HelperFileTest extends HelperTestCase
             $path . '/multipage-landscape.pdf',
             $path . '/multipage-portrait.pdf'
         ];
-        $this->assertEquals($filesystem, HelperConstantsFile::scanFilesystem($path));
-        $this->assertEquals($filesystemWildcard, HelperConstantsFile::scanFilesystem($path, '*.pdf'));
+        $this->assertEquals($filesystem, HelperFile::scanFilesystem($path));
+        $this->assertEquals($filesystemWildcard, HelperFile::scanFilesystem($path, '*.pdf'));
     }
 }